src/main/java/co/kr/steadysprinting/hellospring/repository/MemoryMemberRepository.java
package co.kr.steadysprinting.hellospring.repository;

import co.kr.steadysprinting.hellospring.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;


    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id)); // null 값이 반환 될것을 생각하여 Optional.ofNullable 사용
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name)) // 람다 사용
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values()); // store 의 value 인 Member 모두가 반환 된다.
    }


    public void clearStore(){
        store.clear();
    }
}
